backend/requirements.txt
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.34.0
supabase>=2.15.0
python-dotenv>=1.1.0
pydantic[email]>=2.11.0
httpx>=0.28.0
python-multipart>=0.0.18
